A Desktop Watermarking Application/requirements.txt
Pillow==10.4.0
customtkinter==5.2.2
